Orders

+++ internal/model/order.go
@@ -2,10 +2,10 @@ package model
 
 type (
 	Order struct {
-		Number     int    `json:"number,string"`
-		Status     string `json:"status"`
-		Accrual    int    `json:"accrual,omitempty"`
-		UploadedAt string `json:"uploaded_at"`
+		Number     int     `json:"number,string"`
+		Status     string  `json:"status"`
+		Accrual    float32 `json:"accrual,omitempty"`
+		UploadedAt string  `json:"uploaded_at"`
 	}
 	OrderInPoll struct {
 		Number int
@@ -13,8 +13,8 @@ type (
 		User   int
 	}
 	OrderInAccrual struct {
-		Number  int    `json:"order,string"`
-		Status  string `json:"status"`
-		Accrual int    `json:"accrual,omitempty"`
+		Number  int     `json:"order,string"`
+		Status  string  `json:"status"`
+		Accrual float32 `json:"accrual,omitempty"`
 	}
 )
